+++ src/main/kotlin/io/github/yluttsev/catalogextractor/model/DependencyCoordinate.kt
@@ -1,0 +1,9 @@
+package io.github.yluttsev.catalogextractor.model
+
+data class DependencyCoordinate(
+    val groupId: String,
+    val artifactId: String,
+    val version: String
+) {
+    val module: String get() = "$groupId:$artifactId"
+}